kerosene: Use date-fns instead of moment
