+++ packages/kerosene/package.json
@@ -27,8 +27,8 @@
     "@types/lodash": "^4.14.121",
     "content-type": "^1.0.4",
     "core-js-pure": "^3.1.3",
-    "lodash": "^4.17.11",
-    "moment": "^2.24.0"
+    "date-fns": "^1.30.1",
+    "lodash": "^4.17.11"
   },
   "devDependencies": {},
   "peerDependencies": {},
